Commited some crap!
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@2306 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/environment/Environment.java
@@ -865,7 +865,7 @@ public class Environment implements IEnvironment {
     int ambientAlpha = 0;
     Color ambientColor = Color.WHITE;
     try {
-      ambientAlpha = Integer.parseInt(alphaProp);
+      ambientAlpha = (int)Double.parseDouble(alphaProp);
       ambientColor = Color.decode(colorProp);
     } catch (final NumberFormatException e) {
     }
